requirements.txt
Flask==3.1.0
bcrypt==4.2.1
python-dotenv==1.0.1
requests==2.32.3
Jinja2==3.1.5
Werkzeug==3.1.3
itsdangerous==2.2.0
